requirements.txt
Flask==2.3.3
google-genai==1.2.0
youtube-transcript-api==0.6.2
gunicorn==21.2.0